Dropbox/PC/Desktop/flask app/requirements.txt
streamlit==1.28.0
google-generativeai==0.7.2
python-dotenv==1.0.0
pandas==2.1.0
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
chromadb==0.4.15
sentence-transformers==2.2.2
numpy==1.24.3
pathlib2